+++ src/sreg/models/__init__.py
@@ -1,7 +1,11 @@
 """Data contracts for SREG."""
 
+from sreg.models.agent_tools import AgentTool, AgentToolset
+from sreg.models.benchmark import BenchmarkComparison, BenchmarkResult, BenchmarkStatus
 from sreg.models.case_plan import CasePlan, EvalQuestionPlan
+from sreg.models.code_exec import CodeExecConfig, CodeExecResult, ExecStatus
 from sreg.models.dag_spec import DAGNodeSpec, DAGSpec
+from sreg.models.env_protocol import EnvAction, EnvObservation, EnvStepResult
 from sreg.models.episode import Action, ActionDef, ActionType, Episode, Observation, StepResult
 from sreg.models.research_problem import (
     AvailableAction,
@@ -18,16 +22,27 @@ __all__ = [
     "Action",
     "ActionDef",
     "ActionType",
+    "AgentTool",
+    "AgentToolset",
     "AvailableAction",
+    "BenchmarkComparison",
+    "BenchmarkResult",
+    "BenchmarkStatus",
     "CPD",
     "CasePlan",
+    "CodeExecConfig",
+    "CodeExecResult",
     "DAGNodeSpec",
     "DAGSpec",
     "DataAsset",
     "DifficultyProfile",
     "Edge",
+    "EnvAction",
+    "EnvObservation",
+    "EnvStepResult",
     "Episode",
     "EvalQuestionPlan",
+    "ExecStatus",
     "Node",
     "NodeType",
     "Observation",